refacotr: move values schemas to its own file
- document VALUE_TYPES well
